telco_churn_rate/deployment/requirements.txt
streamlit
pandas
scikit-learn == 1.0.2
tensorflow-cpu